some bug fix for slam6d and fbr
git-svn-id: svn+ssh://svn.code.sf.net/p/slam6d/code/trunk@840 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ include/slam6d/fbr/registration.h
@@ -27,7 +27,7 @@ namespace fbr{
    * @param bErrorIndex error index of registration with bAlign
    */
   class registration{
-    unsigned int minDistance;
+    double minDistance;
     double minError;
     unsigned int minInlier;
     registration_method rMethod;
@@ -73,7 +73,7 @@ namespace fbr{
 
   public:
     registration();
-    registration(unsigned int minD, double minE, unsigned int minI, registration_method method);
+    registration(double minD, double minE, unsigned int minI, registration_method method);
     /**
      * findRegistration : find the transformation matrix between two scans
      * @param fPMap first panorama map which is 3D points coresponding to fist panorama image
@@ -83,7 +83,7 @@ namespace fbr{
      * @param matches matched keypoints from first to second scan
      */
     void findRegistration(cv::Mat fPMap, vector<cv::KeyPoint> fKeypoints, cv::Mat sPMap, vector<cv::KeyPoint> sKeypoints, vector<cv::DMatch> matches);
-    unsigned int getMinDistance();
+    double getMinDistance();
     double getMinError();
     unsigned int getMinInlier();
     registration_method getRegistrationMethod();
